fix: stop truncating transaction latency to int32 microseconds
LatencyRecord narrowed the measured latency on the way into a Sample:

    int latencyMicroseconds = (int) ((endNanosecond - startNanosecond + 500) / 1000);

so anything past 2^31 us -- 35.8 minutes -- silently wrapped. That is harmless
for OLTP, where transactions are milliseconds, and wrong in two different ways
for an analytical stream against a slow engine.

A transaction between 35.8 and 71.6 minutes wraps NEGATIVE, which at least
looks broken: three executions in a TPC-H SF=0.1 run were recorded as -450s,
-514s and -853s. Past 71.6 minutes it wraps back into the POSITIVE range and
looks like an ordinary answer, with nothing in the value to give it away. One
MySQL Q20 reported 611s for an execution that actually took about 4906s; it was
only caught by comparing against the start-time gap, which is reconstructed
from long offsets and does not wrap.

Widened to long. The blast radius is four files and six sites, because the
value is only ever passed along: ThreadBench collects it into an array and an
ArrayList, DistributionStatistics reduces it, ResultWriter prints it.
DistributionStatistics already stored its percentiles as long[], so the int[]
input was the odd one out rather than a deliberate packing choice -- and Sample
has held a long startNanosecond next to the int latency all along, so there was
no per-sample packing argument left to preserve either.

Nothing else in the class changes: the start-time offset was already long, and
the recorded value is still microseconds.

Verified so far: compiles and packages clean under -P cubrid. An end-to-end
check against the query that produced the 2429s wrapped reading is running and
its result is not in this commit.

+++ src/main/java/com/oltpbenchmark/ThreadBench.java
@@ -334,7 +334,7 @@ public class ThreadBench implements Thread.UncaughtExceptionHandler {
       Collections.sort(samples);
 
       // Compute stats on all the latencies
-      int[] latencies = new int[samples.size()];
+      long[] latencies = new long[samples.size()];
       for (int i = 0; i < samples.size(); ++i) {
         latencies[i] = samples.get(i).getLatencyMicrosecond();
       }
@@ -469,7 +469,7 @@ public class ThreadBench implements Thread.UncaughtExceptionHandler {
     private void calculateNext() {
 
       // Collect all samples in the time window
-      ArrayList<Integer> latencies = new ArrayList<>();
+      ArrayList<Long> latencies = new ArrayList<>();
       long endNanoseconds = nextStartNanosecond + (windowSizeSeconds * 1000000000L);
       while (sample != null && sample.getStartNanosecond() < endNanoseconds) {
 
@@ -492,7 +492,7 @@ public class ThreadBench implements Thread.UncaughtExceptionHandler {
 
       nextStartNanosecond = endNanoseconds;
 
-      int[] l = new int[latencies.size()];
+      long[] l = new long[latencies.size()];
       for (int i = 0; i < l.length; ++i) {
         l[i] = latencies.get(i);
       }
